refactor: Move UIView extension to utils
