+++ CardGameApp/CardGameApp/Utils/UIView+FillSuperView.swift
@@ -1,0 +1,19 @@
+//
+//  UIView+FillSuperView.swift
+//  CardGameApp
+//
+//  Created by Chaewan Park on 2020/02/13.
+//  Copyright © 2020 Chaewan Park. All rights reserved.
+//
+
+import UIKit
+
+extension UIView {
+    func fillSuperView() {
+        guard let superview = superview else { return }
+        topAnchor.constraint(equalTo: superview.topAnchor).isActive = true
+        leadingAnchor.constraint(equalTo: superview.leadingAnchor).isActive = true
+        trailingAnchor.constraint(equalTo: superview.trailingAnchor).isActive = true
+        bottomAnchor.constraint(equalTo: superview.bottomAnchor).isActive = true
+    }
+}
