+++ CardGameApp/CardGameApp/Views/PlayModeSelectionView.swift
@@ -70,13 +70,3 @@ class PlayModeSelectionView: UIView {
         }
     }
 }
-
-extension UIView {
-    func fillSuperView() {
-        guard let superview = superview else { return }
-        topAnchor.constraint(equalTo: superview.topAnchor).isActive = true
-        leadingAnchor.constraint(equalTo: superview.leadingAnchor).isActive = true
-        trailingAnchor.constraint(equalTo: superview.trailingAnchor).isActive = true
-        bottomAnchor.constraint(equalTo: superview.bottomAnchor).isActive = true
-    }
-}
